closes #78
