+++ src/components/JUPSettingsMenu/components/AboutUs/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./AboutUs";
